Define remaining rules & extract config classes to own files
