+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/Dependencies.kt
@@ -4,7 +4,7 @@ import io.ktor.client.*
 import io.ktor.client.engine.cio.*
 import io.ktor.client.plugins.*
 import org.regenagcoop.discord.Discord
-import org.regenagcoop.model.ActiveMemberConfig
+import org.regenagcoop.model.config.ActiveMemberConfig
 import org.regenagcoop.tools.GlobalObjectMapper
 import java.io.File
 import kotlin.system.exitProcess
