+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/discord/service/PersistPostsService.kt
@@ -3,7 +3,7 @@ package org.regenagcoop.discord.service
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.regenagcoop.discord.Discord
 import org.regenagcoop.discord.model.UserId
-import org.regenagcoop.model.ActiveMemberConfig
+import org.regenagcoop.model.config.ActiveMemberConfig
 import org.regenagcoop.model.PostHistory
 import java.time.LocalDate
