+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/discord/service/PersistedActivityService.kt
@@ -4,7 +4,7 @@ import io.github.oshai.kotlinlogging.KotlinLogging
 import org.regenagcoop.discord.Discord
 import org.regenagcoop.discord.model.Message
 import org.regenagcoop.discord.model.UserId
-import org.regenagcoop.model.ActiveMemberConfig
+import org.regenagcoop.model.config.ActiveMemberConfig
 import org.regenagcoop.model.RoleChange
 import org.regenagcoop.model.RoleChangeHistory
 import java.time.Instant
