--- gatekeeper-bot/src/main/kotlin/org/regenagcoop/model/KeepRoleConfig.kt
@@ -1,7 +1,0 @@
-package org.regenagcoop.model
-
-data class KeepRoleConfig(
-        val windowSize: Int,
-        /** The minimum number of days the user must post on within the window in order to keep the role */
-        val minPostDays: Int,
-)
