+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/model/config/ActiveMemberConfig.kt
@@ -1,0 +1,29 @@
+package org.regenagcoop.model.config
+
+import org.regenagcoop.discord.model.*
+import java.time.LocalDate
+
+data class ActiveMemberConfig(
+    val guildId: GuildId, // aka "guild id"
+    val excludedUserIds: Set<UserId>, // eg bots
+    val roleConfigs: List<RoleConfig>,
+    val downgradeMessageConfig: DowngradeMessageConfig,
+    val persistenceConfig: PersistenceConfig,
+) {
+    private val maxWindowSize: UInt = roleConfigs.flatMap { cfg ->
+        cfg.paths.map { path -> path.daysToConsider }
+    }.max()
+
+    /**
+     * The earliest date we need to consider when scanning for post history.
+     * Since we can't scan back for reaction history, this is only used for post history
+     */
+    fun computeEarliestScanDate(today: LocalDate): LocalDate {
+        val daysToLookBack = maxWindowSize
+        return if (daysToLookBack == 0u) {
+            today
+        } else {
+            today.minusDays(daysToLookBack.toLong() - 1L)
+        }
+    }
+}
