+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/model/config/Path.kt
@@ -1,0 +1,6 @@
+package org.regenagcoop.model.config
+
+data class Path(
+    val daysToConsider: UInt,
+    val rule: Rule,
+)
