+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/model/config/PersistenceConfig.kt
@@ -1,0 +1,7 @@
+package org.regenagcoop.model.config
+
+import org.regenagcoop.discord.model.ChannelId
+
+data class PersistenceConfig(
+    val channel: ChannelId,
+)
