+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/model/config/RoleConfig.kt
@@ -1,0 +1,9 @@
+package org.regenagcoop.model.config
+
+import org.regenagcoop.discord.model.RoleId
+
+data class RoleConfig(
+    val roleId: RoleId,
+    val paths: List<Path>,
+    val welcomeMessageConfig: WelcomeMessageConfig? = null,
+)
