+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/model/config/WelcomeMessageConfig.kt
@@ -1,0 +1,15 @@
+package org.regenagcoop.model.config
+
+import org.regenagcoop.discord.model.ChannelId
+import org.regenagcoop.discord.model.UserId
+
+data class WelcomeMessageConfig(
+    val channel: ChannelId,
+    val template: String,
+    val userMentionPlaceholder: String = "USER_MENTION",
+) {
+    fun createWelcomeMessage(userId: UserId): String {
+        return template
+            .replace(userMentionPlaceholder, "<@$userId>")
+    }
+}
